Integrated tests into the main code
